Provide a solution to Q2
The header needed to be re-aligned when the window
was re-sized. Use the fle display in the header
and ul elements. Inside the header the content needed
to be given the space-between value. Items were aligned
in the center of the cross axis. Padding of 8px
was applied to help with spacing.

In the ul element, giving a 8px gap separates each
of the tabs and added a margin/padding of 0 to
give a better look.