Provide Solution for Q3
Made alteration to html document by adding a left
and right div elements to help make each section of
the header to separate from each other.

Css document had the header selector given a display
of flex and justify the content to be spaced-between.
The list of the links also were set to display flex,
with a gap of 16px to keep them apart and
set margin/padding to 0. The newly created
left/right classes each had there items-align
set to center, and a gap of 16px applied.