Provide CSS solution for Q4
Use the container selector to justify-content to
the center and a gap of 52px to have each item spaced
out. To align the text of the title it should be centered
as well as provide a margin on the bottom to give it white space.
Each item should only have a max-width of 200px so
they don't grow too large.