Provide html solution for Holy Grail
The main issue in the original html file was the lack
of distinction between sections. The footer and header
were fine, but there needed to be two added divs. One
would contain the elements of the page including the
sidebar and content. The last one was to bundle
the different text elements into one.